Update project form
